area2code {Kormaps}	R Documentation

Seek area from data areacode and returns the code

Description

Seek area from data areacode and returns the code

Usage

area2code(area)

Arguments

area
a string looking for

Value

a code if the area is found, else returns NA

[Package Kormaps version 0.1.0 Index]