areacode {Kormaps}	R Documentation

Korean administrative area name with code

Description

A dataset containing the code and the name of Korean administrative area

Usage

areacode

Format

A data frame with 16 rows and 3 columns

code

code, numeric

name

the full name of Korean administrative area

name2

the abbreviation of name which is more commonly used

[Package Kormaps version 0.1.0 Index]