myleaflet {Kormaps}	R Documentation

Create a Leaflet map widget with multiple map layers and control

Description

Create a Leaflet map widget with multiple map layers and control

Usage

myleaflet(map, choices, popupcol = NULL, palette = "Reds", tile = NULL)

Arguments

map
a data object (currently supported objects are matrices, data frames, and spatial objects from the sp package of classes SpatialPoints, SpatialPointsDataFrame, Polygon, Polygons, SpatialPolygons, SpatialPolygonsDataFrame, Line, Lines, SpatialLines, and SpatialLinesDataFrame)

choices
a character vector specify the column name of map data which being used as a color scale

popupcol
a string specify the column name of map data which being used as a pop-up window

palette
The colors or color function that values will be mapped to

Value

A HTML widget object, on which we can add graphics layers using

[Package Kormaps version 0.1.0 Index]