tbc {Kormaps}	R Documentation

A dataset contains counts of newly developed patients of a disease according to Korean administrative area

Description

A dataset contains counts of newly developed patients of a disease according to Korean administrative area

Usage

tbc

Format

A data frame with 16 rows and 3 columns

<ec>떆<eb>룄蹂<84>

Korean administrative area

Y2001

total number of newly developed patients in the area in 2001

Y2002

total number of newly developed patients in the area in 2002

Y2003

total number of newly developed patients in the area in 2003

Y2004

total number of newly developed patients in the area in 2004

Y2005

total number of newly developed patients in the area in 2005

Y2006

total number of newly developed patients in the area in 2006

Y2007

total number of newly developed patients in the area in 2007

Y2008

total number of newly developed patients in the area in 2008

Y2009

total number of newly developed patients in the area in 2009

Y2010

total number of newly developed patients in the area in 2010

Y2011

total number of newly developed patients in the area in 2011

Y2012

total number of newly developed patients in the area in 2012

Y2013

total number of newly developed patients in the area in 2013

code

code indicating Korean administrative area

[Package Kormaps version 0.1.0 Index]